# Optimize 'tracklabels' plugin